requirements.txt
python-dotenv
langchain
langchain-google-genai
beautifulsoup4
discord
google-genai
langchain-community
ddgs
requests
Pillow